backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0